move combined output functions to FactorGraphModel
